Add_drif_by_transform
Added the option to apply drift as a transformation as an alternative to by each pixel.
Scipy affine_transform v<1.6.0 does not support anything other than mode='constant, so I used skimag's warp function. Warp does not handle cupy arrays. So it is converted to a numpy array, which is not ideal but functional.
Unfortuantely this does not hangle wrapping either and leaves an artifact.
@thomasaarholt Not sure if you have an idea.

Also chagne some drift_vector to drift_angle for consistancy